+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Table of Contents
 
+- [0.2.3](#023---20200518)
 - [0.2.2](#022---20200211)
 - [0.2.1](#021---20200123)
 - [0.2.0](#020---20191218)
 - [0.1.2](#012---20191216)
 - [0.1.1](#011---20191212)
 - [0.1.0](#010---20191212)
+
+##  [0.2.3] - 2020/05/18
+
+- Bump lua-resty-acme to get better error handling.
 
 ##  [0.2.2] - 2020/02/11
 
@@ -34,6 +39,9 @@
 
 - Initial release of ACME plugin for Kong.
 
+[0.2.3]: https://github.com/Kong/kong-plugin-acme/compare/0.2.2...0.2.3
+[0.2.2]: https://github.com/Kong/kong-plugin-acme/compare/0.2.1...0.2.2
+[0.2.1]: https://github.com/Kong/kong-plugin-acme/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/Kong/kong-plugin-acme/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/Kong/kong-plugin-acme/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/Kong/kong-plugin-acme/compare/0.1.0...0.1.1